Refactor and break down trainingCourseModal into smaller components
